ui/requirements.txt
# VisionGuard AI - Debug UI Dependencies

streamlit>=1.31.0
redis>=5.0.0
Pillow>=10.0.0
